chore(docs): moved shared/pages into shared-pages to not confuse eslint rules
