+++ .eslintrc.json
@@ -308,7 +308,7 @@
             }
         },
         {
-            "files": ["*.spec.ts"],
+            "files": ["*.spec.ts", "libs/docs/**/*.ts"],
             "rules": {
                 "jsdoc/require-jsdoc": "off",
                 "grouped-accessor-pairs": "off"
