--- libs/docs/platform/.eslintrc.json
@@ -1,33 +1,0 @@
-{
-    "extends": ["../../../.eslintrc.json"],
-    "ignorePatterns": ["!**/*"],
-    "overrides": [
-        {
-            "files": ["*.ts"],
-            "extends": ["plugin:@nx/angular", "plugin:@angular-eslint/template/process-inline-templates"],
-            "rules": {
-                "@angular-eslint/directive-selector": [
-                    "error",
-                    {
-                        "type": "attribute",
-                        "prefix": "fundamentalNgx",
-                        "style": "camelCase"
-                    }
-                ],
-                "@angular-eslint/component-selector": [
-                    "error",
-                    {
-                        "type": "element",
-                        "prefix": "fundamental-ngx",
-                        "style": "kebab-case"
-                    }
-                ]
-            }
-        },
-        {
-            "files": ["*.html"],
-            "extends": ["plugin:@nx/angular-template"],
-            "rules": {}
-        }
-    ]
-}
