+++ libs/docs/shared-pages/library-doc-shell-page.component.ts
@@ -2,10 +2,12 @@ import { CdkScrollable } from '@angular/cdk/overlay';
 import { NgIf } from '@angular/common';
 import { Component, inject } from '@angular/core';
 import { ActivatedRoute, RouterOutlet } from '@angular/router';
-import { SectionInterface } from '../core-helpers/sections-toolbar/section.interface';
-import { SectionsToolbarComponent } from '../core-helpers/sections-toolbar/sections-toolbar.component';
-import { ToolbarDocsComponent } from '../core-helpers/toolbar/toolbar.component';
-import { DocumentationBaseComponent } from '../documentation-base.component';
+import {
+    DocumentationBaseComponent,
+    SectionInterface,
+    SectionsToolbarComponent,
+    ToolbarDocsComponent
+} from '@fundamental-ngx/docs/shared';
 
 @Component({
     templateUrl: './library-doc-shell-page.component.html',
