+++ libs/docs/shared-pages/new-component-page.component.ts
