+++ libs/docs/shared-pages/tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "extends": "../../../tsconfig.base.json",
+    "files": [],
+    "include": ["index.ts"],
+    "compilerOptions": {
+        "forceConsistentCasingInFileNames": true,
+        "strict": true,
+        "noImplicitReturns": false,
+        "noFallthroughCasesInSwitch": true,
+        "target": "es2022",
+        "outDir": "../../dist/out-tsc",
+        "declaration": true,
+        "declarationMap": true,
+        "inlineSources": true,
+        "types": []
+    },
+    "exclude": ["**/*.test.ts", "**/*.spec.ts", "**/*.e2e-spec.ts"]
+}
