+++ tsconfig.base.json
@@ -304,7 +304,7 @@
             "@fundamental-ngx/docs/platform/wizard-generator": ["libs/docs/platform/wizard-generator/index.ts"],
             "@fundamental-ngx/docs/schema": ["libs/docs/schema/src/index.ts"],
             "@fundamental-ngx/docs/shared": ["libs/docs/shared/src/index.ts"],
-            "@fundamental-ngx/docs/shared/pages": ["libs/docs/shared/src/lib/pages/index.ts"],
+            "@fundamental-ngx/docs/shared-pages": ["libs/docs/shared-pages/index.ts"],
             "@fundamental-ngx/i18n": ["libs/i18n/src/index.ts"],
             "@fundamental-ngx/moment-adapter": ["libs/moment-adapter/src/index.ts"],
             "@fundamental-ngx/nx-plugin": ["libs/nx-plugin/src/index.ts"],
